+++ src/autocomplete/core.ts
@@ -1,0 +1,43 @@
+import { Provider } from '../providers/provider'
+
+export type Direction = {
+  index: number
+  direction: 'forward' | 'backward' | 'still'
+}
+
+export function defaultDirection(): Direction {
+  return { index: 0, direction: 'still' }
+}
+
+export function completionWordIn(
+  editor: CodeMirror.Editor,
+  cursorAtTrigger?: CodeMirror.Position
+) {
+  const cursor = editor.getCursor()
+  const currentLine: string = editor.getLine(cursor.line)
+  const word = currentLine.substring(cursorAtTrigger?.ch || 0, cursor.ch)
+
+  return word
+}
+
+export function managePlaceholders(
+  selectedValue: string,
+  initialCursorIndex: number
+): { normalizedValue: string; newCursorPosition: number } {
+  let normalizedValue: string
+  const placeholder = Provider.placeholder
+  let newCursorPosition = initialCursorIndex
+
+  const placeholderIndex = selectedValue.indexOf(placeholder)
+  if (placeholderIndex > -1) {
+    // TODO: Manage multiple placeholders
+    const placeholderRegex = new RegExp(placeholder, 'g')
+    normalizedValue = selectedValue.replace(placeholderRegex, '')
+    newCursorPosition += placeholderIndex
+  } else {
+    normalizedValue = selectedValue
+    newCursorPosition += selectedValue.length
+  }
+
+  return { normalizedValue, newCursorPosition }
+}