+++ src/providers/provider.ts
@@ -2,6 +2,8 @@ export abstract class Provider {
   abstract readonly category: string
   abstract readonly completions: Array<string>
 
+  static placeholder: string = '#{}'
+
   matchWith(input: string): Completion[] {
     // TODO: Improve filtering with weights
     const suggestions = this.completions.filter(val => val.includes(input))